# [js/webgpu] Optimize NCHW layout for InstanceNormalization