@skottmckay 2023-10-27
Whilst this may help for this model I'm not sure the change is beneficial in general. What stops the Transpose nodes around the InstanceNormalization being optimized away with the existing code? 

Using NCHW for InstanceNormalization with the rest of the nodes assigned to the EP as NHWC would generally mean you're making things worse.

Simplistic example: NCHW data -> Conv -> InstanceNormalization -> Add -> NCHW data

when converted to NHWC layout would first become

NCHW data -> Tranpose(toNHWC) -> Conv(NHWC) -> Transpose(toNCHW) ->Tranpose(toNHWC) -> InstanceNormalization(NHWC) -> Transpose(toNCHW) -> Add

the 2 Tranpose nodes between Conv(NHWC) and InstanceNormalization(NHWC) cancel out, and the last one can be pushed past the Add. That results in one Transpose at the start of the nodes assigned to the EP, and one right at the end. all nodes in the middle are NHWC. 

If you change InstanceNormalization to be NCHW I would think you're creating additional Transpose cost as you're creating more NCHW sections within the nodes assigned to the EP.