@qjia7 2023-10-27
@skottmckay I filed an issue #18128 to trace the transpose optimization issue in the model.

Do you think it's possible that we can dynamically choose NHWC InstanceNormalization or NCHW InstanceNormalization in JESP with NHWC layout? I mean in ORT, when we meet an InstanceNormalization op, if the input is NCHW, we choose NCHW InstanceNormalization. If the input is NHWC, we choose NHWC InstanceNormalization.

I know this PR may be not friendly in your case since it adds two extra transpose. But I am considering if two transposes + NCHW InstanceNormalization is still better than NHWC InstanceNormalization. Does that mean we should always use NCHW InstanceNormalization? Theoretically, NCHW InstanceNormalization is better for gpu memory since we need to calculate mean/meansqure per channel. It makes sure per channel data is continuous. But NHWC InstanceNormalization is not good for gpu memory from the definition.  But I can't say the current NHWC algorithm can't be further optimized. Maybe I need to see whether the current NHWC can be further optimized.

I will remove the change in [layout_transformation.cc](https://github.com/microsoft/onnxruntime/pull/18123/files#diff-ea4394f60cf16c830bbf2b47009268491e5a46824aa3b23c34d889d4bf183980) since other parts are valid so that the left part can be merged. I can do more investigation to see whether we should always use NCHW layout for InstanceNormalization.