@skottmckay 2023-10-31
> Do you think it's possible that we can dynamically choose NHWC InstanceNormalization or NCHW InstanceNormalization in JESP with NHWC layout? I mean in ORT, when we meet an InstanceNormalization op, if the input is NCHW, we choose NCHW InstanceNormalization. If the input is NHWC, we choose NHWC InstanceNormalization.

It think it would be non-trivial. By definition the input to the InstanceNormalization is NCHW to start with as that is what the ONNX spec requires, so there's no simple 'check if input is NHWC or NCHW' that can be done. 

It sounds like you want to traverse the graph possibly in both directions to determine where any existing Transpose nodes are that convert between NHWC and NCHW in order to determine whether you want to convert that particular node to NHWC. That determination probably needs to consider where Transpose nodes can be pushed as well given the layout conversion happens as one step and the transpose optimization happens following that. i.e. the Transpose nodes added to convert to NHWC have not been optimized yet, and to determine which ones will cancel out and/or get blocked would require knowledge of all the internals of the transpose optimizer.